requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
openai>=1.0.0
requests==2.31.0 